editoast: cleanup after the move of all rollings stock structures
