--- editoast/src/schema/rolling_stock/mod.rs
@@ -1,5 +1,0 @@
-pub mod light_rolling_stock;
-
-editoast_common::schemas! {
-    light_rolling_stock::schemas(),
-}
